requirements.txt
yt-dlp
mutagen
python-dotenv
lyricsgenius
